requirements.txt
Flask==3.0.0
opencv-python
pillow